chore(ts): refactor vector source to ts
